Add basic watcher for Pods and Services
This patch introduces K8sApiWatcher, the base class for the more
specific watchers such as K8sPodsWatcher and K8sServicesWatcher.

K8sPodsWatcher and K8sServicesWatcher are introduced at the same time
and they watch the endpoints for each resource and emit the log messages
to watcher.log file for now.

Change-Id: I715c625982f83ac1a0d667c6dfb172660def7855
Signed-off-by: Taku Fukushima <tfukushima@midokura.com>